refactored the code - divided code into utils and used components for image response
