+++ components/Og/Footer.tsx
@@ -1,0 +1,9 @@
+const Footer = () => (
+  <div tw="flex absolute bottom-4 left-[47%] justify-center">
+    <p tw="text-[12px]" style={{ fontWeight: 400 }}>
+      sonata.tips
+    </p>
+  </div>
+);
+
+export default Footer;
