+++ components/Og/Header.tsx
@@ -1,0 +1,10 @@
+const Header = () => (
+  <div tw="flex items-center">
+    <img src="https://i.imgur.com/BHbqO6X.png" width={32} height={32} />
+    <p tw="text-2xl m-0 ml-2 text-[#333536]" style={{ fontWeight: 600 }}>
+      Sonata
+    </p>
+  </div>
+);
+
+export default Header;
