+++ lib/utils.ts
@@ -7,6 +7,7 @@ import { supabaseClient } from '@/lib/supabase/client';
 import findValidEmbed from '@/lib/findValidEmbed';
 import fetchMetadata from '@/lib/fetchMetadata';
 import { CHANNELS } from './consts';
+import getUserByUsername from './neynar/getNeynarUserByUsername';
 
 dayjs.extend(relativeTime);
 
@@ -72,4 +73,40 @@ export function formatPoints(points: any) {
 export function replaceSpecialCharacters(str: any) {
   const specialChars = /[^a-zA-Z0-9]/g;
   return str.replace(specialChars, '_');
+}
+
+export async function getDataForCastOg(username: string, hash: any) {
+  const fullHash: any = await getFullHash(username, hash);
+
+  const encodedUsername = replaceSpecialCharacters(username);
+
+  const userProfile = await getUserByUsername(username);
+  const profilePfp = userProfile?.pfp?.url;
+
+  const { cast, metadata } = await getEmbedAndMetadata(fullHash);
+  const channelData = getChannelData(cast?.channelId);
+
+  const channelLabel = channelData?.label || '/sonata';
+  const channelIcon = channelData?.icon || 'https://i.imgur.com/Xa4LjYA.jpeg';
+
+  const points = formatPoints(cast?.points);
+
+  const soraSemiBold = fetch(new URL('/public/Sora-SemiBold.ttf', import.meta.url)).then((res) =>
+    res.arrayBuffer(),
+  );
+
+  const soraNormal = fetch(new URL('/public/Sora-Regular.ttf', import.meta.url)).then((res) =>
+    res.arrayBuffer(),
+  );
+
+  return {
+    encodedUsername,
+    profilePfp,
+    metadata,
+    channelLabel,
+    channelIcon,
+    points,
+    soraSemiBold,
+    soraNormal,
+  };
 }
