+++ package.json
@@ -32,6 +32,7 @@
     "@vercel/analytics": "^1.2.2",
     "@vercel/og": "^0.6.2",
     "@xmtp/frames-validator": "^0.5.2",
+    "axios": "^1.7.0-beta.2",
     "bufferutil": "^4.0.8",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
